app/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
google-cloud-storage
pydantic
